+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ### UNRELEASED
+- Upgrade react-router@^4.3.1
+- Upgrade react-router-dom@^4.3.1
+- Upgrade react@^16.3.0
+- Upgrade react-dom@^16.3.0
+- Upgrade styled-components@^4.1.3
 
 ### 5.1.3
 - Add url event label to google analysis click event